Update sticky to init on resize and move to own file
